Move all of the files into a subfolder
