+++ macaw/macaw_structural.py
@@ -7,9 +7,9 @@ predicting fluxes through reactions)
 '''
 
 import pandas as pd
-from macaw_utils import add_reaction_equations
+from macaw.macaw_utils import add_reaction_equations
 import itertools as it
-from macaw_utils import flip_reaction
+from macaw.macaw_utils import flip_reaction
 
 def dead_end_test(
     given_model, use_names = False, add_suffixes = False, verbose = 1
